+++ src/types/next-auth.d.ts
@@ -6,5 +6,18 @@ declare module "next-auth" {
       id: string
     } & DefaultSession["user"]
     accessToken?: string
+    firstName?: string
+    lastName?: string
+    email?: string
+    sub?: string
+  }
+
+  interface JWT {
+    accessToken?: string
+    refreshToken?: string
+    expiresAt?: number
+    userId?: string
+    firstName?: string
+    lastName?: string
   }
 }
